requirements.txt
streamlit==1.32.0
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.1
pillow==10.2.0
face-recognition==1.3.0
scikit-learn==1.4.1.post1
matplotlib==3.8.3
retina-face==0.0.15
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
facenet-pytorch==2.5.3
insightface==0.7.3
mediapipe==0.10.9
albumentations==1.4.0
onnxruntime==1.17.1
seaborn==0.12.2
tensorflow==2.15.0